refactor(StationRepository) : 테이블의 모든 데이터를 삭제하는 deleteAll 추가
